konstruktorok

+++ src/Dye.java
@@ -1,7 +1,7 @@
 
 public class Dye extends DyeParent {
 
-    public Dye(String name, int price, int cylinder_volume, int cover) {
+    public Dye(String name, double price, int cylinder_volume, int cover) {
         super(name, price, cylinder_volume, cover);
     }
 }
