+++ src/Lakk.java
@@ -1,6 +1,6 @@
 
 public class Lakk extends DyeParent{
-    public Lakk(String name, int price, int cylinder_volume, int cover) {
+    public Lakk(String name, double price, int cylinder_volume, int cover) {
         super(name, price, cylinder_volume, cover);
     }
 }
