+++ src/Metal.java
@@ -1,7 +1,7 @@
 
 public class Metal extends DyeParent{
 
-    public Metal(String name, int price, int cylinder_volume, int cover) {
+    public Metal(String name, double price, int cylinder_volume, int cover) {
         super(name, price, cylinder_volume, cover);
     }
